Add simple usage examples to README.md
This commit has the goal to provide simple instructions for people that
are new to Flutter (like me) start to work with this package.

This commit adds to README.md simple examples how to connect to a Meteor
backend, how to subscribe to collection updates and how to call Meteor
methods.